Removes deprecated fileio.TextFileSink.
Users should be using textio.WriteToText() transform instead of
fileio.TextFileSink.
